+++ src/serializing.rs
@@ -1,8 +1,4 @@
 //! This module provides serialization functions. Currently contains only markdown serialization.
-//! 
+//!
 
-#[cfg(feature = "markdown")]
 mod md;
-
-#[cfg(feature = "markdown")]
-pub(crate) use md::serialize_md;