+++ src/serializing/md.rs
@@ -1,4 +1,5 @@
 mod constants;
+mod ext;
 mod serializer;
 mod text_utils;